+++ altun.efe/T2/delimiter.hpp
@@ -1,0 +1,14 @@
+#ifndef DELIMITER_HPP
+#define DELIMITER_HPP
+#include <iostream>
+
+namespace altun
+{
+  struct Delimiter
+  {
+    char expected;
+  };
+  std::istream& operator>>(std::istream& in, Delimiter&& exp);
+
+}
+#endif
