+++ altun.efe/T2/streamGuard.cpp
@@ -1,0 +1,12 @@
+#include "streamGuard.hpp"
+
+altun::StreamGuard::StreamGuard(std::basic_ios< char >& s):
+  s_(s),
+  precision_(s.precision()),
+  flags_(s.flags())
+{}
+altun::StreamGuard::~StreamGuard()
+{
+  s_.precision(precision_);
+  s_.flags(flags_);
+}
